api/requirements.txt
# Requerimientos para la API Serverless de Python en Vercel

Flask==3.0.0
tensorflow-cpu==2.18.0
scikit-learn==1.5.0
numpy==1.26.4
joblib==1.4.2
pandas==2.2.0